drop ArrayToYieldDataProviderRector instead, update docs
