+++ config/level/phpunit/phpunit-yield-data-provider.yml
@@ -1,5 +1,4 @@
 services:
-    Rector\PHPUnit\Rector\ArrayToYieldDataProviderRector: ~
-
     Rector\CodingStyle\Rector\ClassMethod\ReturnArrayClassMethodToYieldRector:
-
+        PHPUnit\Framework\TestCase:
+            - '#(provide|dataProvider)*#'
