+++ docs/AllRectorsOverview.md
@@ -1761,7 +1761,7 @@ Null is no more allowed in get_class()
 
 - class: `Rector\Php\Rector\FuncCall\TrailingCommaArgumentsRector`
 
-Adds trailing commas to function and methods calls 
+Adds trailing commas to function and methods calls
 
 ```diff
  calling(
