+++ packages/CodingStyle/tests/Rector/ClassMethod/ReturnArrayClassMethodToYieldRector/Fixture/fixture.php.inc
@@ -22,7 +22,7 @@ use Rector\CodingStyle\Tests\Rector\ClassMethod\ReturnArrayClassMethodToYieldRec
 
 class SomeEventSubscriber implements EventSubscriberInterface
 {
-    public static function getSubscribedEvents()
+    public static function getSubscribedEvents(): \Iterator
     {
         yield 'event' => 'callback';
     }
