+++ packages/CodingStyle/tests/Rector/ClassMethod/ReturnArrayClassMethodToYieldRector/Fixture/fixture4.php.inc
@@ -1,0 +1,36 @@
+<?php
+
+namespace Rector\CodingStyle\Tests\Rector\ClassMethod\ReturnArrayClassMethodToYieldRector\Fixture;
+
+use Rector\CodingStyle\Tests\Rector\ClassMethod\ReturnArrayClassMethodToYieldRector\Source\ParentTestCase;
+
+final class MyTestCase extends ParentTestCase
+{
+    /**
+     * @return mixed[]
+     */
+    public function provideData(): array
+    {
+        return [
+            ['item']
+        ];
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\CodingStyle\Tests\Rector\ClassMethod\ReturnArrayClassMethodToYieldRector\Fixture;
+
+use Rector\CodingStyle\Tests\Rector\ClassMethod\ReturnArrayClassMethodToYieldRector\Source\ParentTestCase;
+
+final class MyTestCase extends ParentTestCase
+{
+    public function provideData(): \Iterator
+    {
+        yield ['item'];
+    }
+}
+
+?>
