+++ packages/CodingStyle/tests/Rector/ClassMethod/ReturnArrayClassMethodToYieldRector/ReturnArrayClassMethodToYieldRectorTest.php
@@ -13,6 +13,9 @@ final class ReturnArrayClassMethodToYieldRectorTest extends AbstractRectorTestCa
     {
         $this->doTestFiles([
             __DIR__ . '/Fixture/fixture.php.inc',
+            __DIR__ . '/Fixture/fixture2.php.inc',
+            __DIR__ . '/Fixture/fixture3.php.inc',
+            __DIR__ . '/Fixture/fixture4.php.inc',
             // @see https://github.com/FriendsOfPHP/PHP-CS-Fixer/commit/53fbbf01ca960f10c88acd0a38689c57bf3ade78
             __DIR__ . '/Fixture/data_provider.php.inc',
         ]);
