--- packages/PHPUnit/tests/Rector/ArrayToYieldDataProviderRector/Fixture/fixture.php.inc
@@ -1,36 +1,0 @@
-<?php
-
-namespace Rector\PHPUnit\Tests\Rector\ArrayToYieldDataProviderRector\Fixture;
-
-use PHPUnit\Framework\TestCase;
-
-final class MyTestCsae extends TestCase
-{
-    /**
-     * @return mixed[]
-     */
-    public function provideData(): array
-    {
-        return [
-            ['item']
-        ];
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\PHPUnit\Tests\Rector\ArrayToYieldDataProviderRector\Fixture;
-
-use PHPUnit\Framework\TestCase;
-
-final class MyTestCsae extends TestCase
-{
-    public function provideData(): \Iterator
-    {
-        yield ['item'];
-    }
-}
-
-?>
