--- packages/PHPUnit/tests/Rector/ArrayToYieldDataProviderRector/Fixture/fixture3.php.inc
@@ -1,31 +1,0 @@
-<?php
-
-namespace Rector\PHPUnit\Tests\Rector\ArrayToYieldDataProviderRector\Fixture;
-
-use PHPUnit\Framework\TestCase;
-
-final class MyTestCase3 extends TestCase
-{
-    public function dataProvider(): array
-    {
-        return [[__DIR__ . '/Fixture/fixture.php.inc', __DIR__ . '/Correct/correct.php.inc']];
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\PHPUnit\Tests\Rector\ArrayToYieldDataProviderRector\Fixture;
-
-use PHPUnit\Framework\TestCase;
-
-final class MyTestCase3 extends TestCase
-{
-    public function dataProvider(): \Iterator
-    {
-        yield [__DIR__ . '/Fixture/fixture.php.inc', __DIR__ . '/Correct/correct.php.inc'];
-    }
-}
-
-?>
